Move old bad smoothing methods to their own module in advance of replacing them.
